ATUverce/backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
groq==0.9.0
dj-database-url==2.1.0